tabControllerおよびCalculatorに関数を分ける
