+++ .vs/99times99/xs/UserPrefs.xml
@@ -2,8 +2,10 @@
   <MonoDevelop.Ide.Workbench ActiveDocument="Assets/Scripts/RandomCalcSceneController.cs">
     <Files>
       <File FileName="Assets/Scripts/SceneTransition.cs" Line="4" Column="35" />
-      <File FileName="Assets/Scripts/RandomCalcSceneController.cs" Line="358" Column="31" />
+      <File FileName="Assets/Scripts/RandomCalcSceneController.cs" Line="113" Column="10" />
       <File FileName="Assets/Scripts/CommentSceneController.cs" />
+      <File FileName="Assets/Scripts/toggleController.cs" Line="6" Column="17" />
+      <File FileName="Assets/Scripts/Calculator.cs" Line="1" Column="1" />
     </Files>
     <Pads>
       <Pad Id="ProjectPad">
