+++ Assembly-CSharp.csproj
@@ -63,10 +63,11 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Assets\Scripts\Calculator.cs" />
     <Compile Include="Assets\Scripts\CommentSceneController.cs" />
     <Compile Include="Assets\Scripts\RandomCalcSceneController.cs" />
     <Compile Include="Assets\Scripts\SceneTransition.cs" />
-    <Compile Include="Assets\Scripts\toggleController.cs" />
+    <Compile Include="Assets\Scripts\tabController.cs" />
     <Reference Include="UnityEngine">
       <HintPath>/Applications/Unity/Hub/Editor/2019.4.5f1/Unity.app/Contents/Managed/UnityEngine/UnityEngine.dll</HintPath>
     </Reference>
