+++ Assets/Scripts/RandomCalcSceneController.cs
@@ -106,15 +106,12 @@ public class RandomCalcSceneController : MonoBehaviour
     public void PushCommentButton()
     {
         /*tab create*/
-        tab_flag_list = Calc_tab_num(get_left_num(),get_right_num());
-        tab_list = new List<GameObject>();
-        for (int i = 0; i < tab_flag_list.Count; i++) {
-            tab_list.Add(SetTabObject(i,tab_flag_list));
-        }
+        tab_flag_list = Calculator.Calc_tab_num(left_num,right_num);
+        tab_list = tabController.get_TabList(tab_flag_list, tab_container);
         tab_list[0].GetComponent<Toggle>().isOn = true;
         tab_list[0].GetComponent<Toggle>().Select();
-        CalcColumnMultiplication(get_left_num(), get_right_num(), page_container.transform.Find("Page0").gameObject);
-        CalcAddDifProduction(get_left_num(), get_right_num(), page_container.transform.Find("Page1").gameObject);
+        Calculator.Calc_ColumnMultiplication(left_num, right_num, page_container.transform.Find("Page0").gameObject);
+        Calculator.Calc_AddDifProduction(left_num, right_num, page_container.transform.Find("Page1").gameObject);
 
         /*move tab panel*/
         StartCoroutine(moveUpCommentPanel(tab_panel,new Vector3(0.0f,-155.0f,0.0f)));
@@ -125,172 +122,6 @@ public class RandomCalcSceneController : MonoBehaviour
         /*move tab panel*/
         StartCoroutine(moveDownCommentPanel(tab_panel, new Vector3(0.0f, -485.0f, 90.0f)));
     }
-
-    public List<int> Calc_tab_num(int left_num, int right_num)
-    {
-        var tab_flag_list = new List<int>();
-        tab_flag_list.Add(1);
-        tab_flag_list.Add((left_num + right_num) % 2 == 1 ? 0 : 1);
-        tab_flag_list.Add(1);
-        return tab_flag_list;
-    }
-
-    public GameObject SetTabObject(int i, List<int> tab_list)
-    {
-
-        Transform tab_trf = tab_container.transform.Find("Tab" + i.ToString());
-
-/*        tab_trf.localPosition = new Vector3(i * 40 - 92, 148, 0);
-        tab_trf.localScale = new Vector3(1, 1, 1);*/
-        if(tab_list[i] == 0){
-            tab_trf.gameObject.GetComponent<Toggle>().interactable = false;
-            tab_trf.Find("Background/text").GetComponent<Text>().color = new Color(0.0f, 0.0f, 0.0f, 0.5f);
-        }
-
-        return tab_trf.gameObject;
-    }
-
-    public void CalcColumnMultiplication(int left_num, int right_num, GameObject clm_page_obj)
-    {
-        int LTenP = left_num / 10;
-        string LTenP_s = LTenP.ToString();
-        if (LTenP == 0)
-        {
-            LTenP_s = "";
-        }
-
-        string LOneP_s = (left_num % 10).ToString();
-
-
-        int RTenP = right_num / 10;
-        string RTenP_s = RTenP.ToString();
-        if (RTenP == 0)
-        {
-            RTenP_s = "";
-        }
-
-        string ROneP_s = (right_num % 10).ToString();
-
-
-        int OHunP = (left_num * (right_num % 10)) / 100;
-        string OHunP_s = OHunP.ToString();
-        if(OHunP == 0)
-        {
-            OHunP_s = "";
-        }
-
-        int OTenP = ((left_num * (right_num % 10)) / 10) % 10;
-        string OTenP_s = OTenP.ToString();
-        if(OHunP ==0 && OTenP == 0)
-        {
-            OTenP_s = "";
-        }
-
-        int OOneP = (left_num * (right_num % 10)) % 10;
-        string OOneP_s = OOneP.ToString();
-
-        int Sub_OTenP = ((left_num % 10) * (right_num % 10)) / 10;
-        string Sub_OTenP_s = Sub_OTenP.ToString();
-        if (Sub_OTenP==0)
-        {
-            Sub_OTenP_s = "";
-        }
-
-
-        int THunP = (left_num * (right_num / 10)) / 100;
-        string THunP_s = THunP.ToString();
-        if (THunP == 0)
-        {
-            THunP_s = "";
-        }
-
-        int TTenP = ((left_num * (right_num / 10)) / 10) % 10;
-        string TTenP_s = TTenP.ToString();
-        if (THunP == 0 && TTenP == 0)
-        {
-            TTenP_s = "";
-        }
-
-        int TOneP = (left_num * (right_num / 10)) % 10;
-        string TOneP_s = TOneP.ToString();
-
-        int Sub_TTenP = ((left_num % 10) * (right_num / 10)) / 10;
-        string Sub_TTenP_s = Sub_TTenP.ToString();
-        if (Sub_TTenP == 0)
-        {
-            Sub_TTenP_s = "";
-        }
-
-        int AThoP = (left_num * right_num) / 1000;
-        string AThoP_s = AThoP.ToString();
-        if (AThoP == 0)
-        {
-            AThoP_s = "";
-        }
-
-        int AHunP = ((left_num * right_num) / 100) % 10;
-        string AHunP_s = AHunP.ToString();
-        if(AThoP==0 && AHunP == 0)
-        {
-            AHunP_s = "";
-        }
-
-        int ATenP = ((left_num * right_num) / 10) % 10;
-        string ATenP_s = ATenP.ToString();
-        if(AThoP==0 && AHunP ==0 && ATenP == 0)
-        {
-            ATenP_s = "";
-        }
-
-        int AOneP = (left_num * right_num) % 10;
-        string AOneP_s = AOneP.ToString();
-
-        int Sub_AHunP = (OTenP + TOneP) / 10;
-        int Sub_AThoP = (Sub_AHunP + OHunP + TTenP) / 10;
-        string Sub_AHunP_s = Sub_AHunP.ToString();
-        string Sub_AThoP_s = Sub_AThoP.ToString();
-        if (Sub_AHunP == 0)
-        {
-            Sub_AHunP_s = "";
-        }
-        if (Sub_AThoP == 0)
-        {
-            Sub_AThoP_s = "";
-        }
-
-
-        clm_page_obj.transform.Find("Objects0/LTenP").GetComponent<Text>().text = LTenP_s;
-        clm_page_obj.transform.Find("Objects0/LOneP").GetComponent<Text>().text = LOneP_s;
-        clm_page_obj.transform.Find("Objects0/RTenP").GetComponent<Text>().text = RTenP_s;
-        clm_page_obj.transform.Find("Objects0/ROneP").GetComponent<Text>().text = ROneP_s;
-        clm_page_obj.transform.Find("Objects1/1HunP").GetComponent<Text>().text = OHunP_s;
-        clm_page_obj.transform.Find("Objects1/1TenP").GetComponent<Text>().text = OTenP_s;
-        clm_page_obj.transform.Find("Objects1/1OneP").GetComponent<Text>().text = OOneP_s;
-        clm_page_obj.transform.Find("Objects2/2HunP").GetComponent<Text>().text = THunP_s;
-        clm_page_obj.transform.Find("Objects2/2TenP").GetComponent<Text>().text = TTenP_s;
-        clm_page_obj.transform.Find("Objects2/2OneP").GetComponent<Text>().text = TOneP_s;
-        clm_page_obj.transform.Find("Objects3/AThoP").GetComponent<Text>().text = AThoP_s;
-        clm_page_obj.transform.Find("Objects3/AHunP").GetComponent<Text>().text = AHunP_s;
-        clm_page_obj.transform.Find("Objects3/ATenP").GetComponent<Text>().text = ATenP_s;
-        clm_page_obj.transform.Find("Objects3/AOneP").GetComponent<Text>().text = AOneP_s;
-        clm_page_obj.transform.Find("Objects1/S1TenP").GetComponent<Text>().text = Sub_OTenP_s;
-        clm_page_obj.transform.Find("Objects2/S2TenP").GetComponent<Text>().text = Sub_TTenP_s;
-        clm_page_obj.transform.Find("Objects3/SAHunP").GetComponent<Text>().text = Sub_AHunP_s;
-        clm_page_obj.transform.Find("Objects3/SAThoP").GetComponent<Text>().text = Sub_AThoP_s;
-
-    }
-
-    public void CalcAddDifProduction(int left_num, int right_num, GameObject clm_page_obj)
-    {
-        int add = (left_num + right_num)/2;
-        int diff = Mathf.Abs(left_num - right_num)/2;
-        clm_page_obj.transform.Find("Objects0/Text1").GetComponent<Text>().text = "① 和÷2：( " + left_num + " + " + right_num + " )÷2 = " +add + "\n② 差÷2：| "+ left_num + " - " + right_num  + " | ÷2 = " + diff.ToString();
-        clm_page_obj.transform.Find("Objects1/Text2").GetComponent<Text>().text = "①^2：" + add + "^2 = " + add*add + "\n②^2：" + diff + "^2 = " + diff*diff;
-        clm_page_obj.transform.Find("Objects2/Text3").GetComponent<Text>().text = "①^2 - ②^2：" + add * add + " - " + diff * diff;
-        clm_page_obj.transform.Find("Objects2/Text4").GetComponent<Text>().text = " = " + (add * add - diff * diff);
-
-    }
-
 
     public IEnumerator moveObject(GameObject obj, Vector3 targetPosition)
     {
