--- Assets/Scripts/toggleController.cs
@@ -1,19 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
-
-public class toggleController : MonoBehaviour
-{
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    public void toggleswitch(GameObject obj)
-    {
-    }
-
-}
